requirements.txt
fastapi==0.104.1
pydantic>==2.5.2
uvicorn>=0.23.0,<0.24.0.post1
PyBoof==0.43.1
Pillow==10.1.0
numpy==1.26.2
python-multipart==0.0.6
